# demo/3-swaggerhub/specs/openapi.yaml
openapi: 3.0.0
info:
  title: TestAPI
  description: Test API pour WS Swagger / OpenAPI
  termsOfService: http://localhost:8080/swagger-ui.html
  contact:
    name: Philippe Bousquet
    url: http://sqli.com
    email: pbousquet@sqli.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0
  version: 0.0.4-oas3
servers:
- url: //localhost:8080/
tags:
- name: goodbye-api
  description: Goodbye API
- name: hello-api
  description: Hello API
paths:
  /api/v1/goodbye:
    get:
      tags:
      - goodbye-api
      summary: goodbye
      operationId: goodbyeUsingGET_1
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoodbyeDto'
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      deprecated: false
  /api/v1/goodbye/{name}:
    get:
      tags:
      - goodbye-api
      summary: goodbye
      operationId: goodbyeUsingGET
      parameters:
      - name: name
        in: path
        description: name
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoodbyeDto'
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      deprecated: false
  /api/v1/hello:
    get:
      tags:
      - hello-api
      summary: Saluer le monde
      operationId: helloUsingGET_1
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HelloDto'
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      deprecated: false
  /api/v1/hello/{name}:
    get:
      tags:
      - hello-api
      summary: Saluer une personne en particulier
      operationId: helloUsingGET
      parameters:
      - name: name
        in: path
        description: Nom de la personne à saluer
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HelloDto'
        "400":
          description: Mauvaise requête, 123 n'est pas une valeurs valide
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      deprecated: false
components:
  schemas:
    GoodbyeDto:
      title: GoodbyeDto
      type: object
      properties:
        message:
          type: string
    HelloDto:
      title: HelloDto
      type: object
      properties:
        message:
          type: string